+++ origa_ui/src/repository/trailbase_repository.rs
@@ -93,7 +93,7 @@ struct UserRow {
     #[serde(default)]
     daily_load: Option<i32>,
     #[serde(default)]
-    known_vocab_hash: Option<i64>,
+    known_vocab_hash: Option<i32>,
 }
 
 impl UserRow {
@@ -137,7 +137,7 @@ impl UserRow {
                     DailyLoad::default()
                 },
             },
-            self.known_vocab_hash.unwrap_or(0) as u64,
+            self.known_vocab_hash.unwrap_or(0) as u32,
         )
     }
 }
@@ -181,7 +181,7 @@ fn user_to_json(user: &User, trailbase_id: &str) -> serde_json::Value {
         "updated_at": user.updated_at().to_rfc3339(),
         "imported_sets": imported_sets_json,
         "daily_load": i32::from(*user.daily_load()),
-        "known_vocab_hash": user.known_vocab_hash() as i64,
+        "known_vocab_hash": user.known_vocab_hash() as i32,
     })
 }
